third_party/blink/web_tests/external/wpt/css/css-transitions/animations/vertical-align-interpolation-expected.txt

This is a testharness.js-based test.
Found 40 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <vertical-align> from [initial] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [initial] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [initial] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [initial] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [initial] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [initial] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [initial] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [initial] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [initial] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [initial] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
[FAIL] CSS Animations: property <vertical-align> from [initial] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "baseline "
[FAIL] CSS Animations: property <vertical-align> from [initial] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "baseline "
[FAIL] CSS Animations: property <vertical-align> from [initial] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "baseline "
[FAIL] CSS Animations: property <vertical-align> from [initial] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] CSS Animations: property <vertical-align> from [initial] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
[FAIL] Web Animations: property <vertical-align> from [initial] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "baseline "
[FAIL] Web Animations: property <vertical-align> from [initial] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "baseline "
[FAIL] Web Animations: property <vertical-align> from [initial] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "baseline "
[FAIL] Web Animations: property <vertical-align> from [initial] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] Web Animations: property <vertical-align> from [initial] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [unset] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [unset] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [unset] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [unset] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] CSS Transitions: property <vertical-align> from [unset] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [unset] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [unset] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [unset] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [unset] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <vertical-align> from [unset] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
[FAIL] CSS Animations: property <vertical-align> from [unset] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "baseline "
[FAIL] CSS Animations: property <vertical-align> from [unset] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "baseline "
[FAIL] CSS Animations: property <vertical-align> from [unset] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "baseline "
[FAIL] CSS Animations: property <vertical-align> from [unset] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] CSS Animations: property <vertical-align> from [unset] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
[FAIL] Web Animations: property <vertical-align> from [unset] to [100px] at (-0.5) should be [-50px]
  assert_equals: expected "- 50px " but got "baseline "
[FAIL] Web Animations: property <vertical-align> from [unset] to [100px] at (0) should be [0px]
  assert_equals: expected "0px " but got "baseline "
[FAIL] Web Animations: property <vertical-align> from [unset] to [100px] at (0.3) should be [30px]
  assert_equals: expected "30px " but got "baseline "
[FAIL] Web Animations: property <vertical-align> from [unset] to [100px] at (0.6) should be [60px]
  assert_equals: expected "60px " but got "100px "
[FAIL] Web Animations: property <vertical-align> from [unset] to [100px] at (1.5) should be [150px]
  assert_equals: expected "150px " but got "100px "
Harness: the test ran to completion.
